requirements.txt
requests==2.32.3
pytest==8.3.5
uvicorn==0.34.0
fastapi==0.115.11
email-validator==2.2.0
python-dotenv==1.0.1
fastapi-pagination==0.12.34
pydantic==2.10.6
psycopg2-binary==2.9.10
sqlalchemy==2.0.39
sqlmodel==0.0.24
faker==37.0.0